fix: fixed issues in admin/non sso user organization settings + overall improvements
