+++ frontend/src/container/OrganizationSettings/__mocks__/OrganizationSettings.ts
@@ -1,0 +1,21 @@
+export const inviteUser = {
+	// eslint-disable-next-line sonarjs/no-duplicate-string
+	email: 'jane@doe.com',
+	name: 'Jane',
+	token: 'fa1f85040480aea0f7055dc17165594a',
+	createdAt: 1711722942,
+	role: 'VIEWER',
+	organization: 'SigNoz Inc. - India',
+};
+
+export const memberUser = {
+	id: '3223a874-5678458745786',
+	name: 'John Doe',
+	email: 'firstUser@test.io',
+	createdAt: 1666357530,
+	profilePictureURL: '',
+	orgId: '1287612376312867312867',
+	groupId: '5678458745786',
+	role: 'ADMIN',
+	organization: 'Test Inc',
+};
